studio-main/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
python-multipart==0.0.6
python-dotenv==1.0.0
psycopg2-binary==2.9.7
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
apscheduler==3.10.4
pydantic[email]==2.5.0
pydantic-settings==2.1.0
bcrypt==4.1.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-jose==3.3.0
jinja2==3.1.2
requests==2.31.0
aiohttp==3.9.1
pillow==10.1.0
email-validator==2.1.0
